Update GameDAO
Added a nameSearch function for specific cases where finding a game by its id is hard.

FInished implementing the delete function by id.

Added a possibleDuplicateDetector function.
This function serves to, when adding a new game, avoid the creation of duplicate instances of a game, returning a boolean based on whether the game_name already exists or not in the database.